playtest16 - The collection update
## ADDITIONS
- New character:
  - Isaac
- New bosses:
  - Jake (@icelingbolt)
  - Plonk (@plinkel)
  - Tana
  - Ashton
  - Walf
  - Phoenix (@.fancyman.)
  - Chip (@chipsicle)
- Altered some bosses
  - The Tutorialist now has a trail of 'shadows', similar to what they have in the tutorial
  - The Tutorialist can now spawn 'servants' that chase the player and explode into projectiles
- Game combat and Game clean music themes by @thugshaker.com
  - With this, there are now a few *music packs* the player can choose from:
    - Goober Shooter 2 (Default)
    - TBOI Mausoleum
    - TBOI Chest
    - The World Looks Red (ULTRAKILL)
- New power item: Pocket Change
  - Uses 5 POWER
  - Spawns one random coin
  - Has a 5% chance to spawn a battery instead
- New item: Hamsa
  - Grants homing to parried projectiles
  - +1 Luck
- New achievement: Item Abuse
  - Reroll a mythic item using The D6
- New enemy voice lines by @plinkel
- Music and SFX volume can now be changes in settings
  - With this, all of the SFX have been normalized and are all .mp3 files
- New player attribute:
  - `parryHoming`: Applies homing to parried projectiles
- Wave 100 bosses now have a special intro cutscene
- New 'Hide popup texts' setting
- Socials for all of the character creators can now be found in the credits menu

## CHANGES
- Nerfed Pyromaniac Enemy: (750HP -> 500HP)
- Buffed Skywalkrs weapon: (2.5 Damage -> 5 Damage) (trueskywalk3r)
- Buffed Sasha Character: (5 Speed -> 6 Speed) (@redjive2)
- Changed Walf Character: Power items can no longer appear in the shop
- Nerfed Cluster Ammo item: (Epic rarity -> Mythic rarity)
- Buffed The D6 power item: (100 POWER cost -> 75 POWER cost)
- Rewrote the collection menu
  - The collection is now a popup instead of a full menu
- All large portrait textures have a white outline
- The player now gets returned to max health when a wave 100 boss spawns
- The `coinChance` bullet attribute can now spawn more valuable coins (higher value coins have less chance of spawning)
- Fighting a boss that is the same as the selected character now give it a random prefix from ['Evil','Fucked up','Twisted']
- Explosions now have a blur effect when fading away
- Projectile size can no longer be less than 5px
- Enemy projectiles with the `fireyAmmo` attribute now applies the `fire` status effect to the player
- Pickups now can only be picked up after being active for more than 25 ticks
- Generic button elements now have a background color and blur effect
- In the challenge select menu, there is now text saying that achievements and collection additions are disabled when having a challenge selected
- Re-added the B Dawgs music credit in the opening cutscene
- The game now saves when modifying a setting slider or dropdown
- Items in the collection now show how many times each item has been bought
- Items that increase max ammo can no longer appear in the shop if the player has infinite ammo
- The JS Bug power item no longer effects bosses
- Rerolling the shop using the D6 now follows item requirements
- The Pepsi and Diet Pepsi power items no longer give `drillTicks` stat ups but instead give `drillChance` stat ups

## BUG FIXES
- Character perfection achievements now get collected after the wave 100 boss is defeated
- Fixed an issue where waves would spawn in the tutorial
- Fixed an issue where the idol enemy would not prevent the player from healing
- Fixed an issue where runs would start with 10 ammo regardless of the players max ammo
- Fixed an issue where all power items would appear in the shop regardless if it was unlocked or not

## OTHER STUFF
- `DeBread.playsound` no longer takes a `volume` arguement
- All bosses have been separated into `bosses` and `minibosses`
  - Bosses spawn in intervals of 100 waves
  - Minibosses spawn in intervals of 25 waves
- Tutorialist textures are now stored in a sub `tutorialist` folder
- New enemy attributes
  - `bossBarTexture (string)`: Replaces the solid color boss bar to a specified texture
  - `noTargetCollision (boolean)`: Disables collision with the enemies target
  - `fireImmunity (boolean)`: Prevents the enemy from catching on fire
  - `noDeathParticles`: Prevents enemy death particles from being created
  - `generalMoveRequirement (function)`: A function that return a Boolean which is ran before every attack, determining if the attack can be performed
  - `onSpawn (function)`: A function that is ran when the enemy spawns
- Boss attacks can now have requirement functions that return a Boolean that determines if an attack can be performed
- The `spawnEnemy` function now returns the enemy element
- dirVels can now be applied to fire

+++ upgrades.js
@@ -85,6 +85,10 @@ const upgrades = [
 
             apply: () => {
                 modifyStat(['ammo','reloadSpeed'], '*=0.9')
+            },
+
+            requirement: () => {
+                return player.stats.ammo.max < Infinity
             }
         },
         bandage: {
@@ -247,6 +251,10 @@ const upgrades = [
             apply: () => {
                 modifyStat(['ammo','max'], '+=3')
                 modifyStat(['ammo','reloadSpeed'], '*=1.1')
+            },
+
+            requirement: () => {
+                return player.stats.ammo.max < Infinity
             }
         },
         medkit: {
@@ -577,6 +585,10 @@ const upgrades = [
                 modifyStat(['ammo','max'], '+=10')
                 modifyStat(['ammo','reloadSpeed'], '*=1.5')
                 modifyStat(['bullet','shotCooldown'], '*=1.25')
+            },
+
+            requirement: () => {
+                return player.stats.ammo.max < Infinity
             }
         },
         drill_ammo: {
@@ -816,6 +828,18 @@ const upgrades = [
                 createNotification('Stats decreased!',`${randomDamageStat[0]} ${randomDamageStat[1]}, ${randomCooldownStat[0]} ${randomCooldownStat[1]}`)
             }
         },
+        hamsa: {
+            name: 'Hamsa',
+            desc: `
+                Grants homing to parried projectiles<br>
+                <cg>+1</cg> Luck
+            `,
+    
+            apply: () => {
+                modifyStat(['player','parryHoming'],'+=1')
+                modifyStat(['shop','luck'],'+=1')
+            }
+        },
     },
     {
         black_hole: {
@@ -834,20 +858,10 @@ const upgrades = [
                 modifyStat(['ammo','reloadSpeed'], '*=1.1')
                 modifyStat(['bullet','speed'], '*=0.5')
                 modifyStat(['bullet','shotCooldown'], '*=1.5')
-            }
-        },
-        cluster_ammo: {
-            name: 'Cluster Ammo',
-            desc: `
-                Bullets split into <cg>+3</cg> when hitting a wall or enemy.<br> <em style="color: grey;">Split bullets deal half damage.</em>
-            `,
-            priceMult: 1.5,
+            },
 
-            apply: () => {
-                modifyStat(['bullet','split'], '+=3')
-                if(player.stats.bullet.splits === 0) {
-                    modifyStat(['bullet','splits'], '+=1')
-                }
+            requirement: () => {
+                return player.stats.ammo.max < Infinity
             }
         },
         electric_ammo: {
@@ -1052,6 +1066,20 @@ const upgrades = [
                 modifyStat(['bullet','shotCooldown'], '*=0.5')
                 player.health = player.stats.player.maxHealth
                 modifyStat(['player','maxHealth'], '+=100')
+            }
+        },
+        cluster_ammo: {
+            name: 'Cluster Ammo',
+            desc: `
+                Bullets split into <cg>+3</cg> when hitting a wall or enemy.<br> <em style="color: grey;">Split bullets deal half damage.</em>
+            `,
+            priceMult: 1.5,
+
+            apply: () => {
+                modifyStat(['bullet','split'], '+=3')
+                if(player.stats.bullet.splits === 0) {
+                    modifyStat(['bullet','splits'], '+=1')
+                }
             }
         },
         the_tophat: {
@@ -1447,6 +1475,23 @@ const powerItems = [
                 player.stats.ammo.current = player.stats.ammo.max
                 DeBread.playSound('audio/reload-long-end.mp3',0.25)
                 updateUI()
+            }
+        },
+        pocket_change: {
+            name: 'Pocket Change',
+            desc: `
+                Uses <cp>5</cp> POWER<br>
+                Spawns one random coin<br>
+                Has a <cg>5%</cg> chance to spawn a battery instead
+                `,
+            charge: 5,
+
+            use: () => {
+                if(DeBread.randomNum(1,20) === 1) {
+                    pickups.battery([...player.centerPos],5)
+                } else {
+                    pickups.coin(getWeightedChance([100,50,20,5,1]),[...player.centerPos],5,1)
+                }
             }
         },
     },
@@ -1891,19 +1936,25 @@ const powerItems = [
             desc: `
                 Uses <cp>75</cp> POWER<br>
                 50% chance to turn all enemies into Lvl 1 Guy<br>
-                50% chance to double all enemies
+                50% chance to double all enemies<br>
+                <br>
+                <em style="color: grey;">Does not work on bosses</em>
             `,
             charge: 75,
 
             use: () => {
                 if(DeBread.randomNum(0,1) === 0) {
                     doge('area').querySelectorAll('.enemy').forEach(enemy => {
-                        spawnEnemy(enemy.data.pos,enemies.guy,1,0)
-                        enemy.data.kill(undefined, true)
+                        if(!enemy.data.data.boss) {
+                            spawnEnemy(enemy.data.pos,enemies.guy,1,0)
+                            enemy.data.kill(undefined, true)
+                        }
                     })
                 } else {
                     doge('area').querySelectorAll('.enemy').forEach(enemy => {
-                        spawnEnemy([enemy.data.pos[0]+1, enemy.data.pos[1]+1],enemy.data.data,enemy.data.level,0)
+                        if(!enemy.data.data.boss) {
+                            spawnEnemy([enemy.data.pos[0]+1, enemy.data.pos[1]+1],enemy.data.data,enemy.data.level,0)
+                        }
                     })
                 }
             }
@@ -2102,11 +2153,12 @@ const powerItems = [
         the_d6: {
             name: 'The D6',
             desc: `
-                Uses <cp>100</cp> POWER<br>
-                Rerolls all items in the shop within their types and rarities. Rerolled items become twice as expensive.
+                Uses <cp>75</cp> POWER<br>
+                Rerolls all items in the shop within their types and rarities.
             `,
-            charge: 100,
+            charge: 75,
             canUseInShop: true,
+            unlockable: true,
             requirement: () => {
                 return player.inPortal
             },
@@ -2123,18 +2175,36 @@ const powerItems = [
                         itemList = elixirs
                     }
                     
-                    const keys = Object.keys(itemList[elem.data.rarity])
-                    const randomKey = keys[DeBread.randomNum(0,keys.length-1)]
-                    const randomItem = itemList[elem.data.rarity][randomKey]
+                    let itemPicked = false
 
-                    items.push({
-                        data: randomItem,
-                        id: randomKey,
-                        rarity: elem.data.rarity,
-                        type: elem.data.type,
-                        cost: DeBread.round(rarities[elem.data.rarity].costBase * (1 + player.wave / 10) * (randomItem.priceMult ?? 1)) * 2
-                    })
+                    while(!itemPicked) {
+                        const keys = Object.keys(itemList[elem.data.rarity])
+                        const randomKey = keys[DeBread.randomNum(0,keys.length-1)]
+                        const randomItem = itemList[elem.data.rarity][randomKey]
+
+                        let meetsRequirement = true
+                        if(randomItem.requirement) {
+                            meetsRequirement = randomItem.requirement()
+                        }
+
+                        if(meetsRequirement) {
+                            items.push({
+                                data: randomItem,
+                                id: randomKey,
+                                rarity: elem.data.rarity,
+                                type: elem.data.type,
+                                cost: DeBread.round(rarities[elem.data.rarity].costBase * (1 + player.wave / 10) * (randomItem.priceMult ?? 1)) * 2
+                            })
+
+                            itemPicked = true
+                        }
+                    }
+
+                    if(elem.data.rarity === 4) {
+                        getAchievement('Item_Abuse')
+                    }
                 })
+
 
                 createShopItems(items)
             }
@@ -2170,6 +2240,7 @@ const powerItems = [
                 -Wisps explode after 500 ticks, dealing up to <cg>100</cg> damage.
             `,
             charge: 75,
+            unlockable: true,
 
             use: () => {
                 const wisp = document.createElement('div')
@@ -2212,6 +2283,7 @@ const powerItems = [
                 Spawns a Tesla Coil on your crosshair, dealing constant electricity damage to random enemies.
             `,
             charge: 75,
+            unlockable: true,
 
             use: () => {
                 const teslaCoil = document.createElement('div')
@@ -2397,7 +2469,7 @@ const powerItems = [
                     bottle.destroy = () => {
                         createParticles([...bottle.pos], 5, 10, [25,50], 250, 'ease-out',{backgroundColor: 'white'})
                         createExplosion([...bottle.pos], 250, 0, 100, true, [[0,0],[0,0],[0,0],0])
-                        DeBread.playSound('audio/pringlesCan.mp3',0.05,DeBread.randomNum(0.95,1.05,5),false)
+                        DeBread.playSound('audio/pringlesCan.mp3',DeBread.randomNum(0.95,1.05,5),false)
                         bottle.remove()
                     }
 
@@ -2444,7 +2516,7 @@ const powerItems = [
                     ['bullet','range','+=5'],
                     ['bullet','critChance','+=5'],
                     ['bullet','critDamageMult','+=0.1'],
-                    ['bullet','drillTicks','+=1'],
+                    ['bullet','drillChance','+=1'],
                     ['bullet','bounces','+=1'],
                     ['melee','damage','+=2'],
                     ['melee','cooldown','*=0.95'],
@@ -2541,7 +2613,7 @@ const powerItems = [
                     ['bullet','range','+=5'],
                     ['bullet','critChance','+=5'],
                     ['bullet','critDamageMult','+=0.1'],
-                    ['bullet','drillTicks','+=1'],
+                    ['bullet','drillChance','+=1'],
                     ['bullet','bounces','+=1'],
                     ['melee','damage','+=2'],
                     ['melee','cooldown','*=0.95'],
@@ -2907,7 +2979,7 @@ function createShopItems(items) {
 
                     let isUnlocked = true
                     if(randomItem.unlockable) {
-                        if(!saveData.items.includes(randomKey)) {
+                        if(!saveData.stats.unlocked[['items','powerItems','elixirs'][itemType]].includes(randomKey)) {
                             isUnlocked = false
                         }
                     }
@@ -3119,13 +3191,38 @@ function createShopItems(items) {
                 updateTutorialGoal()
             }
 
-            if(
-                !saveData.stats.itemsCollected.includes(randomItems[key].id) &&
-                ![2,3].includes(saveData.gameSettings.gamemode) &&
-                saveData.selectedChallenge === 'none'
-            ) {
-                saveData.stats.itemsCollected.push(randomItems[key].id)
+            const typeNames = [
+                'items',
+                'powerItems',
+                'elixirs',
+            ]
+
+            if(![2,3].includes(saveData.gameSettings.gamemode) && saveData.selectedChallenge === 'none') {
+                const collectionList = saveData.stats.collection[typeNames[itemMeta.type]]
+                let collectionHasItem = false
+                for(const key in collectionList) {
+                    if(collectionList[key].name === item.name) {
+                        collectionList[key].count++
+                        collectionHasItem = true
+                        break
+                    }
+                }
+    
+                if(!collectionHasItem) {
+                    collectionList.push({
+                        name: item.name,
+                        count: 1,
+                    })
+                }
             }
+
+            // if(
+            //     !saveData.stats.itemsCollected.includes(randomItems[key].id) &&
+            //     ![2,3].includes(saveData.gameSettings.gamemode) &&
+            //     saveData.selectedChallenge === 'none'
+            // ) {
+            //     saveData.stats.itemsCollected.push(randomItems[key].id)
+            // }
             save()
         }
     }
